Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <math.h>
 int main(){
-    double a, b , c, max_value;
-    scanf("%lf %lf %lf", &a, &b, &c);
+    int a, b , c, max_value;
+    scanf("%d %d %d", &a, &b, &c);
     max_value = fmax(fmax(a, b), c);
-    printf("%.0f", max_value);
+    printf("%d", max_value);
     return 0;
 }
